Create Tables and Add Data changes
Changed primary keys to auto increment in create Tables file as well as some other minor things.

Add data is mostly done but still need to do the contributor roles which has to be done manually and will be kind of tedious

Songs 25-29 Should be the David Bowie songs to use as an example to search by contributor, same thing with Post Malone.